LingTeX-Word: no Single anywhere -- the stage-1 Overflow, diagnosed
RunDocTests compiled (the seven declaration fixes held) and died in its first
section with run-time error 6, Overflow, at a call passing the literal 0 into a
"ByVal wantSpaceAfter As Single" parameter -- no arithmetic anywhere near it.

That is the second time. Stage 1's undiagnosed failure was "s1 = 10" into a
Single, in a different module, and it "went away" when the procedure was split up
-- recorded then as a workaround, with an honest note that it was not a diagnosis.
Two occurrences of one shape on the same Mac build is the diagnosis: converting an
integer into a Single fails on this build, depending on the stack frame. TypeCheck
passed every type including Single because it tests each in a one-line function,
the one kind of frame where it never fails; that is also why "make it smaller"
looked like a fix.

So the engine uses Double throughout: 96 lines across seven modules, every width,
size, gap and position. Double is VBA's native floating type and what the
JavaScript reference has used all along. ReadSingle is ReadDouble. CSng is CDbl.
Line counts are unchanged, so nothing else moved.

Two PROVEN modules are among the seven -- modWrap and modTests -- which is why
RunAllTests must be run again after this: its 79 checks re-prove them. Keeping
Single at that boundary was the alternative, and it would have kept the very type
that fails in play, with the frame-dependence deciding when. The Single in
TypeCheck and MicroDiagnose stays deliberately: it is the reproduction, and the
comment there now says what came of it.

QUICKSTART's "unexplained failure" section is rewritten as the explained one, with
the rule that falls out: error 6 with no arithmetic in sight means look for a
Single first.

vba-lint ALL PASS, parity 78 passed.

+++ LingTeX-Word/src/modWrap.bas
@@ -83,13 +83,13 @@ End Function
 ' firstIndent   indent of the first wrap line
 ' contIndent    indent of every later wrap line
 '-----------------------------------------------------------------------------
-Public Function ComputeWrapLines(colWidths() As Single, noBreakBefore() As Boolean, _
-        ByVal availWidth As Single, ByVal firstIndent As Single, _
-        ByVal contIndent As Single) As Long()
+Public Function ComputeWrapLines(colWidths() As Double, noBreakBefore() As Boolean, _
+        ByVal availWidth As Double, ByVal firstIndent As Double, _
+        ByVal contIndent As Double) As Long()
 
     Dim lines() As Long, nLines As Long
     Dim n As Long, i As Long, j As Long, k As Long
-    Dim cur As Long, curW As Single, budget As Single
+    Dim cur As Long, curW As Double, budget As Double
 
     n = UBound(colWidths) - LBound(colWidths) + 1
     If n <= 0 Then
@@ -168,11 +168,11 @@ End Function
 '
 ' cellWidths is (tierIndex, columnIndex), as produced by modMeasure.
 '-----------------------------------------------------------------------------
-Public Function ColumnWidths(ex As IgtExample, cellWidths() As Single, _
-        ByVal gap As Single) As Single()
+Public Function ColumnWidths(ex As IgtExample, cellWidths() As Double, _
+        ByVal gap As Double) As Double()
 
-    Dim widths() As Single
-    Dim c As Long, t As Long, w As Single
+    Dim widths() As Double
+    Dim c As Long, t As Long, w As Double
 
     If ex.ColCount = 0 Then
         ReDim widths(0 To 0)
